+++ sdk/src/main/java/io/dapr/client/DaprPreviewClient.java
@@ -17,9 +17,13 @@ import io.dapr.client.domain.BulkPublishEntry;
 import io.dapr.client.domain.BulkPublishRequest;
 import io.dapr.client.domain.BulkPublishResponse;
 import io.dapr.client.domain.BulkPublishResponseFailedEntry;
+import io.dapr.client.domain.DeleteJobRequest;
+import io.dapr.client.domain.GetJobRequest;
+import io.dapr.client.domain.GetJobResponse;
 import io.dapr.client.domain.LockRequest;
 import io.dapr.client.domain.QueryStateRequest;
 import io.dapr.client.domain.QueryStateResponse;
+import io.dapr.client.domain.ScheduleJobRequest;
 import io.dapr.client.domain.UnlockRequest;
 import io.dapr.client.domain.UnlockResponseStatus;
 import io.dapr.client.domain.query.Query;
@@ -268,4 +272,36 @@ public interface DaprPreviewClient extends AutoCloseable {
    */
   <T> Subscription subscribeToEvents(
       String pubsubName, String topic, SubscriptionListener<T> listener, TypeRef<T> type);
+
+  /**
+   * Schedules a job using the provided job request details.
+   *
+   * @param scheduleJobRequest The request containing the details of the job to schedule.
+   *                         Must include a name and optional schedule, data, and other related properties.
+   * @return A {@link Mono} that completes when the job scheduling operation is successful or raises an error.
+   * @throws IllegalArgumentException If the request or its required fields like name are null or empty.
+   */
+  public Mono<Void> scheduleJob(ScheduleJobRequest scheduleJobRequest);
+
+  /**
+   * Retrieves details of a specific job.
+   *
+   * @param getJobRequest The request containing the job name for which the details are to be fetched.
+   *      The name property is mandatory.
+   * @return A {@link Mono} that emits the {@link GetJobResponse} containing job details or raises an
+   *      error if the job is not found.
+   * @throws IllegalArgumentException If the request or its required fields like name are null or empty.
+   */
+
+  public Mono<GetJobResponse> getJob(GetJobRequest getJobRequest);
+
+  /**
+   * Deletes a job based on the given request.
+   *
+   * @param deleteJobRequest The request containing the job name to be deleted.
+   *                        The name property is mandatory.
+   * @return A {@link Mono} that completes when the job is successfully deleted or raises an error.
+   * @throws IllegalArgumentException If the request or its required fields like name are null or empty.
+   */
+  public Mono<Void> deleteJob(DeleteJobRequest deleteJobRequest);
 }
